+++ packages/core/src/generators/builder/generator.ts
@@ -87,6 +87,7 @@ const componentMappers: {
 
     const columns = block.children!.map((item) => ({
       blocks: item.children,
+      width: item.component?.options?.width,
     }));
 
     block.component!.options.columns = columns;
@@ -468,16 +469,39 @@ const mapBoundStyles = (bindings: { [key: string]: Binding | undefined }) => {
         };
       }
     } else {
-      bindings[`style.${key}`] = {
-        code: parsed[key],
-        bindingType: 'expression',
-        type: 'single',
-      };
+      if (isGlobalStyle(key)) {
+        console.warn(
+          `The following bound styles are not supported by Builder JSON and have been removed:
+  "${key}": ${parsed[key]}
+          `,
+        );
+      } else {
+        bindings[`style.${key}`] = {
+          code: parsed[key],
+          bindingType: 'expression',
+          type: 'single',
+        };
+      }
     }
   }
 
   delete bindings['style'];
 };
+
+function isGlobalStyle(key: string) {
+  // These are mapped to their respective responsiveStyle and support bindings
+  if (/max-width: (.*?)px/gm.exec(key)) {
+    return false;
+  }
+
+  return (
+    // pseudo class
+    key.startsWith('&:') ||
+    key.startsWith(':') ||
+    // @ rules
+    key.startsWith('@')
+  );
+}
 
 export const blockToBuilder = (
   json: MitosisNode,